backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
Flask-Bcrypt>=1.0
Flask-JWT-Extended>=4.6
Flask-Cors>=4.0
python-dotenv>=1.0

# ── PostgreSQL drivers ─────────────────────────────────────────────────────────
psycopg2-binary>=2.9          # PostgreSQL adapter
pgvector>=0.3.0               # pgvector Python client (Vector column type)

# ── PDF Processing ─────────────────────────────────────────────────────────────
pdfplumber>=0.11.0

# ── Machine Learning & Transformers ───────────────────────────────────────────
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=3.1.0
huggingface-hub>=0.20.0

# ── Vector Search & Embeddings ────────────────────────────────────────────────
faiss-cpu>=1.8.0              # FAISS for in-memory ANN (pre-pgvector)

# ── Scientific Computing ──────────────────────────────────────────────────────
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0

# ── LangChain Framework ───────────────────────────────────────────────────────
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
datasets>=2.20.0

# ── Graph & Network Analysis ──────────────────────────────────────────────────
networkx>=3.0

# ── Visualization ─────────────────────────────────────────────────────────────
matplotlib>=3.7.0
pillow>=9.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
tqdm>=4.65.0
requests>=2.28.0
setuptools>=65.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0
